chore: factor `drop_test` out of `test_common`
